Use os.write() instead of serial.Serial.write()
serial.Serial.write() appears to busy-wait for the serial device to
become writeable. This is very unhelpful! We would rather just throw the
ppp frame away.

Seeing the CPU pegged at 100% when running this on a Raspberry and
sending a large file made me think the byte-stuffing and unstuffing
could be the bottleneck. It isn't. With this change, sending at the
maximum speed of the modem (which admittedly is not very fast!) only
uses about 4% of a single core on a PI 4.

By the time I discovered this, I'd already rewritten the byte-stuffing
code in cython. That's on a separate branch for now.